+++ meta/autoid/autoid_test.go
@@ -541,321 +541,6 @@ func TestNextStep(t *testing.T) {
 	require.Equal(t, int64(30000), nextStep)
 }
 
-func BenchmarkAllocator_Alloc(b *testing.B) {
-	b.StopTimer()
-	store, err := mockstore.NewMockStore()
-	if err != nil {
-		return
-	}
-	defer func() {
-		err := store.Close()
-		if err != nil {
-			b.Fatal(err)
-		}
-	}()
-	dbID := int64(1)
-	tblID := int64(2)
-	err = kv.RunInNewTxn(context.Background(), store, false, func(ctx context.Context, txn kv.Transaction) error {
-		m := meta.NewMeta(txn)
-		err = m.CreateDatabase(&model.DBInfo{ID: dbID, Name: model.NewCIStr("a")})
-		if err != nil {
-			return err
-		}
-		err = m.CreateTableOrView(dbID, &model.TableInfo{ID: tblID, Name: model.NewCIStr("t")})
-		if err != nil {
-			return err
-		}
-		return nil
-	})
-	if err != nil {
-		return
-	}
-	ctx := context.Background()
-	alloc := autoid.NewAllocator(store, 1, false, autoid.RowIDAllocType)
-	b.StartTimer()
-	for i := 0; i < b.N; i++ {
-		_, _, err := alloc.Alloc(ctx, 2, 1, 1, 1)
-		if err != nil {
-			b.Fatal(err)
-		}
-	}
-}
-
-func BenchmarkAllocator_SequenceAlloc(b *testing.B) {
-	b.StopTimer()
-	store, err := mockstore.NewMockStore()
-	if err != nil {
-		return
-	}
-	defer func() {
-		err := store.Close()
-		if err != nil {
-			b.Fatal(err)
-		}
-	}()
-	var seq *model.SequenceInfo
-	var sequenceBase int64
-	err = kv.RunInNewTxn(context.Background(), store, false, func(ctx context.Context, txn kv.Transaction) error {
-		m := meta.NewMeta(txn)
-		err = m.CreateDatabase(&model.DBInfo{ID: 1, Name: model.NewCIStr("a")})
-		if err != nil {
-			return err
-		}
-		seq = &model.SequenceInfo{
-			Start:      1,
-			Cycle:      true,
-			Cache:      false,
-			MinValue:   -10,
-			MaxValue:   math.MaxInt64,
-			Increment:  2,
-			CacheValue: 2000000,
-		}
-		seqTable := &model.TableInfo{
-			ID:       1,
-			Name:     model.NewCIStr("seq"),
-			Sequence: seq,
-		}
-		sequenceBase = seq.Start - 1
-		err = m.CreateSequenceAndSetSeqValue(1, seqTable, sequenceBase)
-		return err
-	})
-	if err != nil {
-		return
-	}
-	alloc := autoid.NewSequenceAllocator(store, 1, seq)
-	b.StartTimer()
-	for i := 0; i < b.N; i++ {
-		_, _, _, err := alloc.AllocSeqCache(1)
-		if err != nil {
-			fmt.Println("err")
-		}
-	}
-}
-
-func BenchmarkAllocator_Seek(b *testing.B) {
-	base := int64(21421948021)
-	offset := int64(-351354365326)
-	increment := int64(3)
-	b.StartTimer()
-	for i := 0; i < b.N; i++ {
-		_, err := autoid.CalcSequenceBatchSize(base, 3, increment, offset, math.MinInt64, math.MaxInt64)
-		if err != nil {
-			b.Fatal(err)
-		}
-	}
-}
-
-func TestSequenceAutoid(t *testing.T) {
-	store, err := mockstore.NewMockStore()
-	require.NoError(t, err)
-	defer func() {
-		err := store.Close()
-		require.NoError(t, err)
-	}()
-
-	var seq *model.SequenceInfo
-	var sequenceBase int64
-	err = kv.RunInNewTxn(context.Background(), store, false, func(ctx context.Context, txn kv.Transaction) error {
-		m := meta.NewMeta(txn)
-		err = m.CreateDatabase(&model.DBInfo{ID: 1, Name: model.NewCIStr("a")})
-		require.NoError(t, err)
-		seq = &model.SequenceInfo{
-			Start:      1,
-			Cycle:      true,
-			Cache:      true,
-			MinValue:   -10,
-			MaxValue:   10,
-			Increment:  2,
-			CacheValue: 3,
-		}
-		seqTable := &model.TableInfo{
-			ID:       1,
-			Name:     model.NewCIStr("seq"),
-			Sequence: seq,
-		}
-		sequenceBase = seq.Start - 1
-		err = m.CreateSequenceAndSetSeqValue(1, seqTable, sequenceBase)
-		require.NoError(t, err)
-		return nil
-	})
-	require.NoError(t, err)
-
-	alloc := autoid.NewSequenceAllocator(store, 1, seq)
-	require.NotNil(t, alloc)
-
-	// allocate sequence cache.
-	base, end, round, err := alloc.AllocSeqCache(1)
-	require.NoError(t, err)
-	require.Equal(t, int64(0), base)
-	require.Equal(t, int64(5), end)
-	require.Equal(t, int64(0), round)
-
-	// test the sequence batch size.
-	offset := seq.Start
-	size, err := autoid.CalcSequenceBatchSize(sequenceBase, seq.CacheValue, seq.Increment, offset, seq.MinValue, seq.MaxValue)
-	require.NoError(t, err)
-	require.Equal(t, end-base, size)
-
-	// simulate the next value allocation.
-	nextVal, ok := autoid.SeekToFirstSequenceValue(base, seq.Increment, offset, base, end)
-	require.True(t, ok)
-	require.Equal(t, int64(1), nextVal)
-	base = nextVal
-
-	nextVal, ok = autoid.SeekToFirstSequenceValue(base, seq.Increment, offset, base, end)
-	require.True(t, ok)
-	require.Equal(t, int64(3), nextVal)
-	base = nextVal
-
-	nextVal, ok = autoid.SeekToFirstSequenceValue(base, seq.Increment, offset, base, end)
-	require.True(t, ok)
-	require.Equal(t, int64(5), nextVal)
-
-	base, end, round, err = alloc.AllocSeqCache(1)
-	require.NoError(t, err)
-	require.Equal(t, int64(5), base)
-	require.Equal(t, int64(10), end)
-	require.Equal(t, int64(0), round)
-
-	// test the sequence batch size.
-	size, err = autoid.CalcSequenceBatchSize(sequenceBase, seq.CacheValue, seq.Increment, offset, seq.MinValue, seq.MaxValue)
-	require.NoError(t, err)
-	require.Equal(t, end-base, size)
-
-	nextVal, ok = autoid.SeekToFirstSequenceValue(base, seq.Increment, offset, base, end)
-	require.True(t, ok)
-	require.Equal(t, int64(7), nextVal)
-	base = nextVal
-
-	nextVal, ok = autoid.SeekToFirstSequenceValue(base, seq.Increment, offset, base, end)
-	require.True(t, ok)
-	require.Equal(t, int64(9), nextVal)
-	base = nextVal
-
-	_, ok = autoid.SeekToFirstSequenceValue(base, seq.Increment, offset, base, end)
-	// the rest in cache in not enough for next value.
-	require.False(t, ok)
-
-	base, end, round, err = alloc.AllocSeqCache(1)
-	require.NoError(t, err)
-	require.Equal(t, int64(-11), base)
-	require.Equal(t, int64(-6), end)
-	// the round is already in cycle.
-	require.Equal(t, int64(1), round)
-
-	// test the sequence batch size.
-	size, err = autoid.CalcSequenceBatchSize(sequenceBase, seq.CacheValue, seq.Increment, offset, seq.MinValue, seq.MaxValue)
-	require.NoError(t, err)
-	require.Equal(t, end-base, size)
-
-	offset = seq.MinValue
-	nextVal, ok = autoid.SeekToFirstSequenceValue(base, seq.Increment, offset, base, end)
-	require.True(t, ok)
-	require.Equal(t, int64(-10), nextVal)
-	base = nextVal
-
-	nextVal, ok = autoid.SeekToFirstSequenceValue(base, seq.Increment, offset, base, end)
-	require.True(t, ok)
-	require.Equal(t, int64(-8), nextVal)
-	base = nextVal
-
-	nextVal, ok = autoid.SeekToFirstSequenceValue(base, seq.Increment, offset, base, end)
-	require.True(t, ok)
-	require.Equal(t, int64(-6), nextVal)
-	base = nextVal
-
-	_, ok = autoid.SeekToFirstSequenceValue(base, seq.Increment, offset, base, end)
-	// the cache is already empty.
-	require.False(t, ok)
-}
-
-func TestConcurrentAllocSequence(t *testing.T) {
-	store, err := mockstore.NewMockStore()
-	require.NoError(t, err)
-	defer func() {
-		err := store.Close()
-		require.NoError(t, err)
-	}()
-
-	var seq *model.SequenceInfo
-	var sequenceBase int64
-	err = kv.RunInNewTxn(context.Background(), store, false, func(ctx context.Context, txn kv.Transaction) error {
-		m := meta.NewMeta(txn)
-		err1 := m.CreateDatabase(&model.DBInfo{ID: 2, Name: model.NewCIStr("a")})
-		require.NoError(t, err1)
-		seq = &model.SequenceInfo{
-			Start:      100,
-			Cycle:      false,
-			Cache:      true,
-			MinValue:   -100,
-			MaxValue:   100,
-			Increment:  -2,
-			CacheValue: 3,
-		}
-		seqTable := &model.TableInfo{
-			ID:       2,
-			Name:     model.NewCIStr("seq"),
-			Sequence: seq,
-		}
-		if seq.Increment >= 0 {
-			sequenceBase = seq.Start - 1
-		} else {
-			sequenceBase = seq.Start + 1
-		}
-		err1 = m.CreateSequenceAndSetSeqValue(2, seqTable, sequenceBase)
-		require.NoError(t, err1)
-		return nil
-	})
-	require.NoError(t, err)
-
-	var mu sync.Mutex
-	wg := sync.WaitGroup{}
-	m := map[int64]struct{}{}
-	count := 10
-	errCh := make(chan error, count)
-
-	allocSequence := func() {
-		alloc := autoid.NewSequenceAllocator(store, 2, seq)
-		for j := 0; j < 3; j++ {
-			base, end, _, err1 := alloc.AllocSeqCache(2)
-			if err1 != nil {
-				errCh <- err1
-				break
-			}
-
-			errFlag := false
-			mu.Lock()
-			// sequence is negative-growth here.
-			for i := base - 1; i >= end; i-- {
-				if _, ok := m[i]; ok {
-					errCh <- fmt.Errorf("duplicate id:%v", i)
-					errFlag = true
-					mu.Unlock()
-					break
-				}
-				m[i] = struct{}{}
-			}
-			if errFlag {
-				break
-			}
-			mu.Unlock()
-		}
-	}
-	for i := 0; i < count; i++ {
-		wg.Add(1)
-		go func(num int) {
-			time.Sleep(time.Duration(num%10) * time.Microsecond)
-			allocSequence()
-			wg.Done()
-		}(i)
-	}
-	wg.Wait()
-
-	close(errCh)
-	err = <-errCh
-	require.NoError(t, err)
-}
-
 // Fix a computation logic bug in allocator computation.
 func TestAllocComputationIssue(t *testing.T) {
 	require.NoError(t, failpoint.Enable("github.com/pingcap/tidb/meta/autoid/mockAutoIDCustomize", `return(true)`))
